requirements.txt
Gooey==1.0.8.1
pytaglib==1.5.0
requests==2.27.1
